Résolution bug animation après cooldown de dégravite
