+++ Assets/Scenes/SampleScene.unity
@@ -349,6 +349,7 @@ GameObject:
   - component: {fileID: 292136888}
   - component: {fileID: 292136890}
   - component: {fileID: 292136891}
+  - component: {fileID: 292136892}
   m_Layer: 0
   m_Name: Player
   m_TagString: Player
@@ -398,6 +399,7 @@ MonoBehaviour:
   rb: {fileID: 292136883}
   moveSpeed: 5
   note: {fileID: 1763629811}
+  playerAnimator: {fileID: 292136889}
 --- !u!4 &292136886
 Transform:
   m_ObjectHideFlags: 0
@@ -509,7 +511,7 @@ MonoBehaviour:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 292136881}
-  m_Enabled: 1
+  m_Enabled: 0
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 6bf892a94105890419d106feae68632e, type: 3}
   m_Name: 
@@ -571,6 +573,19 @@ PolygonCollider2D:
       - {x: 0.1370352, y: 0.124291785}
       - {x: 0.08877477, y: 0.21661651}
   m_UseDelaunayMesh: 0
+--- !u!114 &292136892
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 292136881}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: aa7a38c95f0191944962354f5e421a68, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  gameManager: {fileID: 0}
 --- !u!1 &519420028
 GameObject:
   m_ObjectHideFlags: 0
