+++ Assets/Scripts/OpenDoor.cs
@@ -1,0 +1,23 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class OpenDoor : MonoBehaviour
+{
+    public GameManager gameManager;
+    // Start is called before the first frame update
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        if (gameManager.notesCollected == gameManager.numberOfNotes)
+        {
+            gameManager.OpenDoor();
+        }
+    }
+
+}
