+++ Assets/Scripts/OpenDoor.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6bf892a94105890419d106feae68632e
+guid: aa7a38c95f0191944962354f5e421a68
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
